+++ src/glorper_line.rs
@@ -1,5 +1,5 @@
 pub struct GlorperLine{
 
-    pub start: (f32, f32),
-    pub end : (f32, f32),
+    pub start: (f64, f64),
+    pub end : (f64, f64),
 }
